+++ src/index.js
@@ -1,5 +1,12 @@
-export * from './components/editor'
+// Main oyez-editor components
+export * from './components/oyez-editor'
 export * from './components/editor-block'
+
+// Data and state management
+export * from './model/editor-state'
+
+// Editors
 export * from './components/editor-quill'
 export * from './components/editor-image'
-export * from './model/editor-state'
+export * from './components/editor-tui'
+export * from './components/editor-codemirror'